+++ MapboxSearch/ui/src/main/java/com/mapbox/search/ui/view/place/SearchPlace.kt
@@ -202,7 +202,6 @@ public class SearchPlace(
          *
          * @param searchResult Search result as base data for place creation.
          * @param responseInfo Search response and request information.
-         * @param coordinate Geojson point with place coordinates.
          * @param distanceMeters Distance in meters to the given search place.
          *
          * @return Search place instance
@@ -212,7 +211,6 @@ public class SearchPlace(
         public fun createFromSearchResult(
             searchResult: SearchResult,
             responseInfo: ResponseInfo,
-            coordinate: Point,
             distanceMeters: Double? = searchResult.distanceMeters
         ): SearchPlace {
             return SearchPlace(
@@ -222,7 +220,7 @@ public class SearchPlace(
                 address = searchResult.address,
                 resultTypes = searchResult.types,
                 record = (searchResult as? IndexableRecordSearchResult)?.record,
-                coordinate = coordinate,
+                coordinate = searchResult.coordinate,
                 routablePoints = searchResult.routablePoints,
                 categories = searchResult.categories,
                 makiIcon = searchResult.makiIcon,
@@ -267,7 +265,6 @@ public class SearchPlace(
          * Creates a new search place instance from [IndexableRecord] and geojson point.
          *
          * @param record A record describing geo place.
-         * @param coordinate Geojson point with place coordinates.
          * @param distanceMeters Distance in meters to the given search place.
          *
          * @return Search place instance
@@ -275,7 +272,6 @@ public class SearchPlace(
         @JvmStatic
         public fun createFromIndexableRecord(
             record: IndexableRecord,
-            coordinate: Point,
             distanceMeters: Double?
         ): SearchPlace {
             val feedback = when (record) {
@@ -290,7 +286,7 @@ public class SearchPlace(
                 address = record.address,
                 resultTypes = listOf(record.type),
                 record = record,
-                coordinate = coordinate,
+                coordinate = record.coordinate,
                 routablePoints = record.routablePoints,
                 categories = record.categories,
                 makiIcon = record.makiIcon,